Recognize 'openat' as a valid open() syscall
In our current test environment, the Python open() function results in
the 'openat' syscall (rather than 'open').

The test suite has been updated to only recognize 'openat' in the
specific cases. It will take a little more work to support both 'open'
and 'openat' because the position of the `pathname` argument differs.