Conditionally format values for matching
We previously always string-ified syscall values in preparation for
filtering them with 'match' checkers. This isn't the behavior we want
for 'byte' and 'byterray' values, however, because we'd end up matching
against their `__str__` representations (e.g. `bytearray(b'abc')`).

We now only format the value when it isn't one of (bytes, bytearray,
str), and we require the matcher to be bytes- or string-aware. (This is
something we could consider revisiting by attempting to decode bytes to
strings, to make matching more natural, but I don't know if that level
of implicit conversion is really needed for such a low-level library.)